requirements.txt
Flask==2.3.3
requests==2.31.0
pandas>=2.0.0
plotly==5.17.0
numpy>=1.24.0
beautifulsoup4==4.12.2
lxml==4.9.3
openai>=1.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7 